Stop adding a Content-Type header to add boundary
Instead, add the boundary to the Content-Type header if it exists
and instead of check that it is a multipart message by the
`message_type`, assume it's multipart if it has children.
This way there is no need to set `message_type`, it's not assumed
to be `multipart/mixed` if it has a Content-Type and the Content-Type
can have other parameters.