Don't let headers contain newlines, fix fn emit_folded()
The problem was:

MIME_LINE_LENGTH is 78. If a line is 78 chars long and ends with a
space, then the previous implementation would insert a line break after
the space (as the line lenghth of 78 is reached).

Then, it would forget about it and go on with the next char, which is
the existing line break.

Therefore we have two line breaks in a row, i.e. an empty line.

The easiest way I found to fix this was, not parse the input character
by character but split it using the stdlib and then work with the
individual words.